requirements.txt
streamlit
langchain
langchain-openai
langchain-community
chromadb
requests
python-dotenv
dashscope
openai
tiktoken